requirements.txt
pyserial
git+https://github.com/adafruit/Adafruit_Nokia_LCD.git
Pillow